security: fix code scanning findings (path injection, format strings, secret wipe, TLS, dynamic dispatch)
Address CodeQL code scanning alerts:

- serve_wizard.py: validate static-file paths against WIZARD_DIR after
  realpath resolution to prevent path traversal (py/path-injection).
- mock_ota_server.py: require TLS 1.2+ on the mock OTA server, refusing
  legacy TLS 1.0/1.1 (py/insecure-protocol).
- mesh_network.cpp: wipe X25519 shared secrets with the optimizer-proof
  secure_wipe() instead of memset, which the compiler may elide
  (cpp/memset-may-be-deleted).
- rf_presence.cpp: cast size_t arguments to unsigned for %u format
  specifiers in conformance log lines (cpp/wrong-type-format-argument).
- spa/app.js: dispatch hash routes only via own, callable properties so a
  user-controlled hash cannot reach inherited members
  (js/unvalidated-dynamic-method-call).

+++ firmware/projects/canary-wap/arduino/canary_wap/rf_presence.cpp
@@ -1265,7 +1265,8 @@ bool conformance_check_token_rotation() {
   }
   if (!tokens_cleared) {
     health_logging::logf(health_logging::LEVEL_ERROR, health_logging::CAT_RF,
-      "Conformance: token map not cleared (had %u, now %u)", old_token_count, s_token_count);
+      "Conformance: token map not cleared (had %u, now %u)", old_token_count,
+      (unsigned)s_token_count);
   }
 
   return epoch_incremented && tokens_differ && tokens_cleared;
@@ -1310,7 +1311,8 @@ bool conformance_check_secure_wipe() {
   for (size_t i = 0; i < sizeof(test_buffer); i++) {
     if (test_buffer[i] != 0) {
       health_logging::logf(health_logging::LEVEL_ERROR, health_logging::CAT_RF,
-        "Conformance: secure_wipe failed at byte %u (value 0x%02X)", i, test_buffer[i]);
+        "Conformance: secure_wipe failed at byte %u (value 0x%02X)", (unsigned)i,
+        test_buffer[i]);
       return false;
     }
   }
